requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
